# public/openapi.yaml
openapi: 3.1.0
info:
  title: DevTrack API
  description: REST API for DevTrack - an open-source developer productivity dashboard
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://devtrack.vercel.app
    description: Production server
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
...
paths:
...
  /api/user/settings:
    get:
      tags: [User]
      summary: Get user settings
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User settings
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
    put:
      tags: [User]
      summary: Update user settings
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Settings updated
        '401':
          description: Unauthorized
...
  /api/user/github-accounts/{githubId}:
    delete:
      tags: [User]
      summary: Unlink a GitHub account
      security:
        - cookieAuth: []
      parameters:
        - name: githubId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account unlinked
        '401':
          description: Unauthorized
        '404':
          description: Account not found

  # NOTIFICATIONS ROUTES
...
  /api/leaderboard/rebuild:
    post:
      tags: [Leaderboard]
      summary: Full leaderboard rebuild
      security:
        - bearerAuth: []
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token with LEADERBOARD_REBUILD_TOKEN
      responses:
        '200':
          description: Leaderboard rebuild started
        '401':
          description: Invalid or missing rebuild token

  # PUBLIC PROFILE ROUTES
...
  /api/ai/roast:
    post:
      tags: [AI]
      summary: Generate AI roast or hype of coding style
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: AI-generated content
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
        '500':
          description: AI service unavailable (GROQ_API_KEY not configured)
...